feat(tags): batch ticket tag edits into one request
Tag editing now runs as a session: chips update optimistically while the
picker is open, and the whole edit lands as a single update_tags call when
it closes, logging one activity instead of one per tag.

The colour step moves into its own TagColorPicker component, and the tag
components are grouped under ticket-agent/tag/.
