+++ desk/src/components/ticket-agent/tag/TagChip.vue
@@ -5,10 +5,7 @@
     size="lg"
     class="min-w-0 max-w-40 gap-1.5 transition-colors duration-150 hover:bg-surface-gray-2"
   >
-    <span
-      class="size-2 shrink-0 rounded-full"
-      :style="{ backgroundColor: color }"
-    />
+    <span class="size-2 shrink-0 rounded-full" :class="color" />
     <span :title="tag" class="min-w-0 truncate text-ink-gray-7">{{ tag }}</span>
   </Badge>
 </template>
