+++ desk/src/composables/useTags.ts
@@ -1,19 +1,39 @@
-import { call, createListResource } from "frappe-ui";
-import { toValue, type MaybeRefOrGetter } from "vue";
+import { createListResource } from "frappe-ui";
 
 export interface Tag {
   name: string;
   color?: string;
 }
 
-/**
- * Tag CRUD for a single document. `tagListResource` is the shared master list
- * of helpdesk tags. `add` routes through the document's own `add_tag` method,
- * which creates the helpdesk Tag (with colour) if missing and links it in a
- * single call; `remove` unlinks via the core endpoint. The doctype must expose
- * an `add_tag(label, color)` whitelisted method (HD Ticket does).
- */
-export function useTags(doctype: string, docname: MaybeRefOrGetter<string>) {
+const DEFAULT_COLOR_CLASS = "bg-surface-gray-5";
+
+// keys are stored on the Tag doc (Select field); values must be full literal
+// Tailwind classes - the JIT scanner can't see `bg-surface-${name}`
+export const TAG_COLORS: Record<string, string> = {
+  Gray: DEFAULT_COLOR_CLASS,
+  Black: "bg-surface-gray-10",
+  Blue: "bg-surface-blue-6",
+  Green: "bg-surface-green-6",
+  Red: "bg-surface-red-6",
+  Pink: "bg-surface-pink-6",
+  Orange: "bg-surface-orange-6",
+  Amber: "bg-surface-amber-6",
+  Yellow: "bg-surface-yellow-6",
+  Cyan: "bg-surface-cyan-6",
+  Teal: "bg-surface-teal-6",
+  Violet: "bg-surface-violet-6",
+  Purple: "bg-surface-purple-6",
+};
+export const TAG_COLOR_NAMES = Object.keys(TAG_COLORS);
+
+/** Dot class for a Tag's stored colour, falling back to Gray. */
+export function colorToken(color?: string): string {
+  return TAG_COLORS[color ?? ""] ?? DEFAULT_COLOR_CLASS;
+}
+
+/** Shared master list of helpdesk tags. Per-document tag changes go through
+ * the document resource's whitelisted `update_tags` method (see useTicket). */
+export function useTags() {
   const tagListResource = createListResource({
     doctype: "Tag",
     fields: ["name", "color"],
@@ -24,22 +44,5 @@ export function useTags(doctype: string, docname: MaybeRefOrGetter<string>) {
     auto: true,
   });
 
-  function add(label: string, color = "Gray") {
-    return call("run_doc_method", {
-      dt: doctype,
-      dn: toValue(docname),
-      method: "add_tag",
-      args: { label, color },
-    });
-  }
-
-  function remove(tag: string) {
-    return call("frappe.desk.doctype.tag.tag.remove_tag", {
-      tag,
-      dt: doctype,
-      dn: toValue(docname),
-    });
-  }
-
-  return { tagListResource, add, remove };
+  return { tagListResource };
 }
